Moving advanced table spec tests location
